code(Closes #1, Fixes #2, Fixes #3, Closes #4)
CAUTION!!! I didn't commit faulty code, but i'm aware of the 6 errors IntelliJ is reporting. The IDE reports a missing fxml controller, but it was defined in the mainView class. The application compiles + runs flawlessly. I think these errors are bugs of the IDE, THERE IS NOTHING WRONG WITH THE CODE!

Finished implementing the main game UI. Everything works, looks good and scales nicely. All FXID fields and actions are documented and implemented in the mainWindowController class.

mainWindowController: There are still some commands to forward to the model, but since these commands are not implemented, i wrote some todo's. Nevertheless, the class is mostly done. Renamed the class from UIController to mainWindowController.

Fixed some minor things in view, and renamed it to mainView.

PlayerColor enum class was created. Maybe it could be embedded in Player class.