code (Closes #6) Implemented Question Window UI
An easy and elegant question window was implemented. The handling is very easy, and it's simple to get a boolean answer from the user.
To create an informational window, you will need to instance it permanently. You can simply create one by using:
InformationalWindow thisQuestion = new InformationalWindow("The title","The text you want to tell the user);
And you can retrieve a boolean value of the user's input by thisQuestion.getAnswer; - It's simple!