code (closes #9, fixes #1, fixes #2) Implemented PlayerEntryWindow and made things neater
I'm happy to announce that the GUI is almost completely functional. If you choose to start a new game the program will ask your player count, initials, throws error windows if the entries are invalid, adds the players to the GUI and field, initializes values and makes the little player color balls visible.
Also enhanced the MainWindowController to initialize the game, i marked two code chunks that should be refactored into logic with a TODO.
Made all-around improvement to the entire GUI system. I think we are ready to implement the rest. I made super sure you will know how to interface with the controller once you browsed through the javadoc.